# Improve performance for command and selection changes