--- a/.changeset/metal-pugs-explain.md
+++ b/.changeset/metal-pugs-explain.md
@@ -1,0 +1,5 @@
+---
+"@guardian/prosemirror-elements": patch
+---
+
+Improve performance in large articles containing many elements