--- a/src/plugin/helpers/__tests__/fixtures.ts
+++ b/src/plugin/helpers/__tests__/fixtures.ts
@@ -43,6 +43,7 @@ export const schema: Schema = new Schema({
 });
 
 export const {
+  doc,
   example,
   example__caption,
   example__html,