--- a/src/plugin/helpers/constants.ts
+++ b/src/plugin/helpers/constants.ts
@@ -2,6 +2,7 @@
 // renderers that must be able to uniquely identify children – for example, in
 
 import { PluginKey } from "prosemirror-state";
+import { PluginState } from "../plugin";
 
 // React, where a `key` attribute is expected in arrays of ReactNodes.
 export const RepeaterFieldMapIDKey = "__ID";
@@ -9,7 +10,7 @@ export const RepeaterFieldMapIDKey = "__ID";
 // A placeholder value for a dropdown option that represents no selection.
 export const undefinedDropdownValue = "none-selected";
 
-export const pluginKey = new PluginKey("prosemirror_elements");
+export const pluginKey = new PluginKey<PluginState>("prosemirror_elements");
 
 export const buttonWidth = 32;
 export const actionSpacing = buttonWidth;