@jonathonherbert 2025-01-23
> Note: have not tested in Composer as (I assume) that requires bumping the prosemirror-elements dependency in the flexible-content repo and I'm not sure how to achieve that locally before this PR merges into main

There's a bit in the main readme to facilitate testing in Composer that I mention above (very briefly — easy to miss!), but I'm quite confident in the test suite for this library, and I did some testing in Composer myself for this PR per the GIFs. I think we're fine to merge and test the new version in Composer when it's available.